+++ flow-typed/npm/express_v4.x.x.js
@@ -1,13 +1,11 @@
-// flow-typed signature: 2c4468fea4b4e486e30ea5b47fbfb75b
-// flow-typed version: c7f4cf7a4d/express_v4.x.x/flow_>=v0.93.x
-
-import * as http from 'http';
-import type { Socket } from 'net';
+// flow-typed signature: 7d0da7d41860b91769189ed541e2ba17
+// flow-typed version: c6154227d1/express_v4.x.x/flow_>=v0.104.x
 
 declare type express$RouterOptions = {
   caseSensitive?: boolean,
   mergeParams?: boolean,
-  strict?: boolean
+  strict?: boolean,
+  ...
 };
 
 declare class express$RequestResponseBase {
@@ -15,15 +13,13 @@ declare class express$RequestResponseBase {
   get(field: string): string | void;
 }
 
-declare type express$RequestParams = {
-  [param: string]: string
-}
+declare type express$RequestParams = { [param: string]: string, ... }
 
 declare class express$Request extends http$IncomingMessage mixins express$RequestResponseBase {
   baseUrl: string;
   body: mixed;
-  cookies: {[cookie: string]: string};
-  connection: Socket;
+  cookies: { [cookie: string]: string, ... };
+  connection: net$Socket;
   fresh: boolean;
   hostname: string;
   ip: string;
@@ -33,10 +29,10 @@ declare class express$Request extends http$IncomingMessage mixins express$Reques
   params: express$RequestParams;
   path: string;
   protocol: 'https' | 'http';
-  query: {[name: string]: string | Array<string>};
+  query: { [name: string]: string | Array<string>, ... };
   route: string;
   secure: boolean;
-  signedCookies: {[signedCookie: string]: string};
+  signedCookies: { [signedCookie: string]: string, ... };
   stale: boolean;
   subdomains: Array<string>;
   xhr: boolean;
@@ -46,7 +42,7 @@ declare class express$Request extends http$IncomingMessage mixins express$Reques
   acceptsEncodings(...encoding: Array<string>): string | false;
   acceptsLanguages(...lang: Array<string>): string | false;
   header(field: string): string | void;
-  is(type: string): boolean;
+  is(type: string): string | false;
   param(name: string, defaultValue?: string): string | void;
 }
 
@@ -58,7 +54,8 @@ declare type express$CookieOptions = {
   maxAge?: number,
   path?: string,
   secure?: boolean,
-  signed?: boolean
+  signed?: boolean,
+  ...
 };
 
 declare type express$Path = string | RegExp;
@@ -69,33 +66,34 @@ declare type express$SendFileOptions = {
   maxAge?: number,
   root?: string,
   lastModified?: boolean,
-  headers?: {[name: string]: string},
-  dotfiles?: 'allow' | 'deny' | 'ignore'
+  headers?: { [name: string]: string, ... },
+  dotfiles?: 'allow' | 'deny' | 'ignore',
+  ...
 };
 
 declare class express$Response extends http$ServerResponse mixins express$RequestResponseBase {
   headersSent: boolean;
-  locals: {[name: string]: mixed};
+  locals: { [name: string]: mixed, ... };
   append(field: string, value?: string): this;
   attachment(filename?: string): this;
   cookie(name: string, value: string, options?: express$CookieOptions): this;
   clearCookie(name: string, options?: express$CookieOptions): this;
   download(path: string, filename?: string, callback?: (err?: ?Error) => void): this;
-  format(typesObject: {[type: string]: Function}): this;
+  format(typesObject: { [type: string]: Function, ... }): this;
   json(body?: mixed): this;
   jsonp(body?: mixed): this;
-  links(links: {[name: string]: string}): this;
+  links(links: { [name: string]: string, ... }): this;
   location(path: string): this;
   redirect(url: string, ...args: Array<void>): this;
   redirect(status: number, url: string, ...args: Array<void>): this;
-  render(view: string, locals?: {[name: string]: mixed}, callback?: express$RenderCallback): this;
+  render(view: string, locals?: { [name: string]: mixed, ... }, callback?: express$RenderCallback): this;
   send(body?: mixed): this;
   sendFile(path: string, options?: express$SendFileOptions, callback?: (err?: ?Error) => mixed): this;
   sendStatus(statusCode: number): this;
   header(field: string, value?: string): this;
-  header(headers: {[name: string]: string}): this;
+  header(headers: { [name: string]: string, ... }): this;
   set(field: string, value?: string|string[]): this;
-  set(headers: {[name: string]: string}): this;
+  set(headers: { [name: string]: string, ... }): this;
   status(statusCode: number): this;
   type(type: string): this;
   vary(field: string): this;
@@ -166,13 +164,13 @@ declare class express$Router extends express$Route {
 
 declare class express$Application extends express$Router mixins events$EventEmitter {
   constructor(): void;
-  locals: {[name: string]: mixed};
+  locals: { [name: string]: mixed, ... };
   mountpath: string;
-  listen(port: number, hostname?: string, backlog?: number, callback?: (err?: ?Error) => mixed): ?http.Server;
-  listen(port: number, hostname?: string, callback?: (err?: ?Error) => mixed): ?http.Server;
-  listen(port: number, callback?: (err?: ?Error) => mixed): ?http.Server;
-  listen(path: string, callback?: (err?: ?Error) => mixed): ?http.Server;
-  listen(handle: Object, callback?: (err?: ?Error) => mixed): ?http.Server;
+  listen(port: number, hostname?: string, backlog?: number, callback?: (err?: ?Error) => mixed): ?http$Server;
+  listen(port: number, hostname?: string, callback?: (err?: ?Error) => mixed): ?http$Server;
+  listen(port: number, callback?: (err?: ?Error) => mixed): ?http$Server;
+  listen(path: string, callback?: (err?: ?Error) => mixed): ?http$Server;
+  listen(handle: Object, callback?: (err?: ?Error) => mixed): ?http$Server;
   disable(name: string): void;
   disabled(name: string): boolean;
   enable(name: string): express$Application;
@@ -183,7 +181,7 @@ declare class express$Application extends express$Router mixins events$EventEmit
    */
   //   get(name: string): mixed;
   set(name: string, value: mixed): mixed;
-  render(name: string, optionsOrFunction: {[name: string]: mixed}, callback: express$RenderCallback): void;
+  render(name: string, optionsOrFunction: { [name: string]: mixed, ... }, callback: express$RenderCallback): void;
   handle(req: http$IncomingMessage<>, res: http$ServerResponse, next?: ?express$NextFunction): void;
   // callable signature is not inherited
   (req: http$IncomingMessage<>, res: http$ServerResponse, next?: ?express$NextFunction): void;
@@ -200,8 +198,12 @@ declare module 'express' {
   declare export type $Application = express$Application;
 
   declare module.exports: {
-    (): express$Application, // If you try to call like a function, it will use this signature
-    static: (root: string, options?: Object) => express$Middleware, // `static` property on the function
-    Router: typeof express$Router, // `Router` property on the function
+    // If you try to call like a function, it will use this signature
+    (): express$Application,
+    // `static` property on the function
+    static: (root: string, options?: Object) => express$Middleware,
+    // `Router` property on the function
+    Router: typeof express$Router,
+    ...
   };
 }