+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 
 ## 1.1.0
 
- 
+ * Core: web-app Plugins can now additionally have handlers for upgrade requests (WebSocket support) and for unload
+ * Core: The _Middleware_ tab in the admin UI shows now the actual order of the stack (until now the order was just calculated) 
 
 ## 1.0.94